backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.15.0
pillow==10.1.0
scikit-image==0.22.0
scikit-learn==1.3.2
imgaug==0.4.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
aiofiles==23.2.1
